SCRUM-47: JSON Plan Schema Loading

+++ src/resources/world_time.rs
@@ -3,14 +3,14 @@ use brahe::Epoch;
 
 #[derive(Resource, Debug)]
 pub struct WorldTime {
-    pub multiplier: f64,
+    pub multiplier: u32,
     pub epoch: Epoch,
 }
 
 impl Default for WorldTime {
     fn default() -> Self {
         Self {
-            multiplier: 1.0,
+            multiplier: 1,
             epoch: Epoch::now(),
         }
     }
